Move SeedSource to the runner module
